toastmasters/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
fonttools==4.38.0
html5lib==1.1
Markdown==3.4.1
mongoengine==0.24.2
pdfkit==1.0.0
Pillow==9.4.0
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pymongo==4.3.3
pyphen==0.13.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
weasyprint==57.2
webencodings==0.5.1
wkhtmltopdf==0.2
zopfli==0.2.2
